Add files via upload
This research on illegal migrants coming to European countries was conducted during my internship at de Volkskrant. The data was taken on EURODAC, 
the European Union fingerprint database for identifying asylum seekers and irregular border-crossers. The aim of this research was to identify how many immigrants are not respecting the Dublin Treaty, given the current changes applied to the treaty.